Initial check-in for unbundling the massive resource files from the DLL's into their own files: 1. Change the project files to not include them as embedded resources. 2. Add a new library to the unit test folder to get the local paths to them for unit testing (I still need to make it smarter to actually get the local path using MSBuild magic instead of the current hard-coded value). 3. Add functionality to NodeWithJavaBase to download them from Azure storage. 4. Modify the PowerShell deployment script to upload them to Azure storage. I still have lots of work to do, but right now Cassandra service works end-to-end with unbundled resources.

+++ Tests/Microsoft.Experimental.Azure.Cassandra.Tests/CassandraNodeRunnerTest.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿using Microsoft.Experimental.Azure.CommonTestUtilities;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -16,7 +17,10 @@ namespace Microsoft.Experimental.Azure.Cassandra.Tests
 		public void EndToEndTest()
 		{
 			var tempDirectory = @"C:\CassandraTestOutput";
-			Directory.Delete(tempDirectory, recursive: true);
+			if (Directory.Exists(tempDirectory))
+			{
+				Directory.Delete(tempDirectory, recursive: true);
+			}
 			var config = new CassandraConfig(
 				clusterName: "Test cluster",
 				clusterNodes: new[] { "127.0.0.1" },
@@ -24,6 +28,7 @@ namespace Microsoft.Experimental.Azure.Cassandra.Tests
 				commitLogDirectory: Path.Combine(tempDirectory, "commitlog"),
 				savedCachesDirectory: Path.Combine(tempDirectory, "savedcaches"));
 			var runner = new CassandraNodeRunner(
+				resourceFileDirectory: ResourcePaths.CassandraResourcesPath,
 				jarsDirectory: Path.Combine(tempDirectory, "jars"),
 				javaHome: @"C:\Program Files\Java\jdk1.7.0_21",
 				logsDirctory: Path.Combine(tempDirectory, "logs"),
